requirements.txt
asgiref==3.4.1
Django==3.2.7
django-cors-headers==3.9.0
django-environ==0.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
Pillow==8.3.2
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.3
sqlparse==0.4.2
